build connection

+++ routes/messages.js
@@ -4,10 +4,9 @@ const pool = require("../db");
 // get all user messages
 router.get("/", async (req, res) => {
   try {
-    const { userId } = req.params;
     const allMessages = await pool.query(
-      "SELECT * FROM messages WHERE user_id = $1",
-      [userId]
+      "SELECT users.username, messages.message_id, messages.message_text FROM users LEFT JOIN messages ON users.user_id = messages.user_id WHERE users.user_id = $1",
+      [req.user.id]
     );
     res.json(allMessages.rows);
   } catch (err) {
@@ -35,10 +34,10 @@ router.post("/", async (req, res) => {
   try {
     const { messageText } = req.body;
     const newMessage = await pool.query(
-      "INSERT INTO messages (message_text) VALUES ($1) RETURNING *",
-      [messageText]
+      "INSERT INTO messages (user_id, message_text) VALUES ($1, $2) RETURNING *",
+      [req.user.id, messageText]
     );
-    res.json(newMessage.rows);
+    res.json(newMessage.rows[0]);
   } catch (err) {
     console.error(err.message);
     res.status(500).json("server error");
@@ -51,13 +50,13 @@ router.put("/:id", async (req, res) => {
     const { messageId } = req.params;
     const { message } = req.body;
     const updateMessage = await pool.query(
-      "UPDATE messages SET message_text = $1 WHERE message_id=$2",
-      [message, messageId]
+      "UPDATE messages SET message_text = $1 WHERE message_id=$2 AND user_id = $3",
+      [message, messageId, req.user.id]
     );
-    res.json("updated");
+    res.json(" message updated");
   } catch (err) {
     console.error(err.message);
-    res.status(500).json("serevr error");
+    res.status(500).json("server error");
   }
 });
 
@@ -68,8 +67,8 @@ router.delete("/:id", async (req, res) => {
     const { messageId } = req.params;
     const { message } = req.body;
     const deleteMessage = await pool.query(
-      "DELETE FROM messages WHERE message_id = $1",
-      [messageId]
+      "DELETE FROM messages WHERE message_id = $1 AND user_id = $2 RETURNING *",
+      [messageId, req.user.id]
     );
     res.json("deleted");
   } catch (err) {
